Create event_model.py file and move transient classes there
The file event_model.py has been created to hold the classes defining
the event model. Transient classes previously in dst_io.py have been
moved there.
